+++ HISTORY.md
@@ -1,6 +1,7 @@
 2.x
 ===
 
+  * Accept `RegExp` object for `hostname`
   * Remove old invocation of `server.onvhost`
 
 1.0.0 / 2014-03-05